Move dashboard types to shared types
Extract DashboardData, DashboardUsage and FormatLimitResult into src/types and remove duplicated type definitions from routes/utils. Update components and routes to import these types from '@/types' (using type-only imports in components). Tighten formatTokenLimit parameter types to use Picks of AIUsage/AIMonthlyUsage and simplify its call sites by removing now-unnecessary feature fields. General cleanup to centralize type definitions and reduce coupling between modules.
